grid-based movement added
Grid-based movement works now.

The door to change scenes doesn't work anymore because the player just controls a normal sprite now.
Also there are clippings and overlappings of tiles over the player that have to be fixed.
The Grid-Engine plugin also needs a few seconds to load and multiplied the loading time in general.

The code was ordered so that GameScene.ts now looks a lot cleaner.